+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/main.js
@@ -1,5 +1,5 @@
 import { renderComments } from "./renderComments.js";
-import { commentLocal } from "./infoLocalStorage.js";
+import { commentLocal } from "./module/infoLocalStorage.js";
 import { mostrarMiPerfil } from "./youProfile.js";
 import { deleteCommentary } from "./funtionDeleteComment.js";
 commentLocal.comments.map((comentario) => renderComments(comentario));
