+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/module/infoLocalStorage.js
@@ -1,5 +1,5 @@
-import { comentarios } from "./comments.js";
-import localStorage from "./module/localStorage.js";
+import { comentarios } from "../comments.js";
+import localStorage from "./localStorage.js";
 if (localStorage.get("comentarios") == null) {
   localStorage.save("comentarios", comentarios);
 }
