+++ NEWBIE/TO-DO interactive-comments-section-main/JavaScript/youProfile.js
@@ -1,6 +1,6 @@
 import { sectionAddComentary } from "./card/commentary.js";
 import action from "./data/acciones.js";
-import { commentLocal } from "./infoLocalStorage.js";
+import { commentLocal } from "./module/infoLocalStorage.js";
 export const mostrarMiPerfil = () => {
   const contentHome = document.getElementById("content-home");
   const youProfile = sectionAddComentary(
